requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
Pillow==10.0.1
python-dotenv==1.0.0
marshmallow==3.20.1
